Moved event listeners from timelines to home
